Major changes to support rails 2.3.x; new version 2.0.0
It was decided that rails 2.3.x would be supported using the line of 2.x
releases of this gem. This is the first release of this line: 2.0.0.
This supports rails 2.3.x and has a cucumber-based test suite that has a
dummy rails 2.3.14 app.

All work on this line of versions will be done in the 'rails2' branch.

Future versions will similarly have their own branch as well. Rails 3.x
support will be present in the 3.x series of this gem.